Refactor file and class names.
