+++ kedro_kubeflow/context_helper.py
@@ -37,10 +37,11 @@ class ContextHelper(object):
     @lru_cache()
     def kfp_client(self):
         if self.config.is_vertex_ai_pipelines:
-            from .vertexaiclient import KubeflowClient
+            from .vertex_ai.client import VertexAIPipelinesClient
+            return VertexAIPipelinesClient(self.config, self.project_name, self.context)
         else:
             from .kfpclient import KubeflowClient
-        return KubeflowClient(self.config, self.project_name, self.context)
+            return KubeflowClient(self.config, self.project_name, self.context)
 
     @staticmethod
     def init(metadata, env):
